Changes on models through all layers

+++ domain/src/main/java/xyz/gonzapico/domain/model/DomainModelLocation.java
@@ -5,22 +5,22 @@ package xyz.gonzapico.domain.model;
  */
 
 public class DomainModelLocation {
-  private long latitude;
-  private long longitude;
+  private double latitude;
+  private double longitude;
 
-  public long getLatitude() {
+  public double getLatitude() {
     return latitude;
   }
 
-  public void setLatitude(long latitude) {
+  public void setLatitude(double latitude) {
     this.latitude = latitude;
   }
 
-  public long getLongitude() {
+  public double getLongitude() {
     return longitude;
   }
 
-  public void setLongitude(long longitude) {
+  public void setLongitude(double longitude) {
     this.longitude = longitude;
   }
 }
